Implement optional type of profile definitions.
Profile definitions may be defined using either DEFVAR or DEFPARAMETER,
depending on the value of *PROFILE-TYPE*. When :VARIABLE, the profile
will be defined using a DEFVAR form; when :PARAMETER the profile will be
defined using DEFPARAMETER instead to better support interactive development.

*PROFILE-TYPE* defaults to :VARIABLE and is not exported to avoid breaking
existing code. A redefined profile still needs to be activated using the
SWITCH-PROFILE function.